requirements.txt
openai
langchain
langchain-openai
langchain-community
streamlit
python-dotenv
PyPDF2

#for installation of local packages
-e .